Introducing engine-runner-common
