+++ engine/runner/src/main/java/org/enso/runner/Main.java
@@ -12,7 +12,6 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.UUID;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Function;
 import java.util.stream.Collectors;
@@ -28,9 +27,6 @@ import org.enso.common.LanguageInfo;
 import org.enso.distribution.DistributionManager;
 import org.enso.distribution.Environment;
 import org.enso.editions.DefaultEdition;
-import org.enso.languageserver.boot.LanguageServerConfig;
-import org.enso.languageserver.boot.ProfilingConfig;
-import org.enso.languageserver.boot.StartupConfig;
 import org.enso.libraryupload.LibraryUploader.UploadFailedError;
 import org.enso.pkg.Contact;
 import org.enso.pkg.PackageManager;
@@ -40,6 +36,7 @@ import org.enso.polyglot.Module;
 import org.enso.polyglot.PolyglotContext;
 import org.enso.profiling.sampler.NoopSampler;
 import org.enso.profiling.sampler.OutputStreamSampler;
+import org.enso.runner.common.LanguageServerApi;
 import org.enso.version.VersionDescription;
 import org.graalvm.polyglot.PolyglotException;
 import org.graalvm.polyglot.PolyglotException.StackFrame;
@@ -1173,84 +1170,6 @@ public final class Main {
     }
   }
 
-  /**
-   * Handles `--server` CLI option
-   *
-   * @param line a CLI line
-   * @param logLevel log level to set for the engine runtime
-   */
-  private void runLanguageServer(CommandLine line, Level logLevel) {
-    try {
-      var config = parseServerOptions(line);
-      LanguageServerApp.run(config, logLevel, line.hasOption(Main.DAEMONIZE_OPTION));
-      throw exitSuccess();
-    } catch (WrongOption e) {
-      System.err.println(e.getMessage());
-      throw exitFail();
-    }
-  }
-
-  private static LanguageServerConfig parseServerOptions(CommandLine line) throws WrongOption {
-    UUID rootId;
-    try {
-      var id = line.getOptionValue(ROOT_ID_OPTION);
-      if (id == null) {
-        throw new WrongOption("Root id must be provided");
-      }
-      rootId = UUID.fromString(id);
-    } catch (IllegalArgumentException e) {
-      throw new WrongOption("Root must be UUID");
-    }
-    var rootPath = line.getOptionValue(ROOT_PATH_OPTION);
-    if (rootPath == null) {
-      throw new WrongOption("Root path must be provided");
-    }
-    var interfac = line.getOptionValue(INTERFACE_OPTION, "127.0.0.1");
-    int rpcPort;
-    try {
-      rpcPort = Integer.parseInt(line.getOptionValue(RPC_PORT_OPTION, "8080"));
-    } catch (NumberFormatException e) {
-      throw new WrongOption("Port must be integer");
-    }
-    int dataPort;
-    try {
-      dataPort = Integer.parseInt(line.getOptionValue(DATA_PORT_OPTION, "8081"));
-    } catch (NumberFormatException e) {
-      throw new WrongOption("Port must be integer");
-    }
-    Integer secureRpcPort;
-    try {
-      var port = line.getOptionValue(SECURE_RPC_PORT_OPTION);
-      secureRpcPort = port == null ? null : Integer.valueOf(port);
-    } catch (NumberFormatException e) {
-      throw new WrongOption("Port must be integer");
-    }
-    Integer secureDataPort;
-    try {
-      var port = line.getOptionValue(SECURE_DATA_PORT_OPTION);
-      secureDataPort = port == null ? null : Integer.valueOf(port);
-    } catch (NumberFormatException e) {
-      throw new WrongOption("Port must be integer");
-    }
-    var profilingConfig = parseProfilingConfig(line);
-    var graalVMUpdater = line.hasOption(SKIP_GRAALVM_UPDATER);
-
-    var config =
-        new LanguageServerConfig(
-            interfac,
-            rpcPort,
-            scala.Option.apply(secureRpcPort),
-            dataPort,
-            scala.Option.apply(secureDataPort),
-            rootId,
-            rootPath,
-            profilingConfig,
-            new StartupConfig(graalVMUpdater),
-            "language-server",
-            ExecutionContext.global());
-    return config;
-  }
-
   private static ProfilingConfig parseProfilingConfig(CommandLine line) throws WrongOption {
     Path profilingPath = null;
     try {
@@ -1293,7 +1212,7 @@ public final class Main {
   }
 
   @SuppressWarnings("unchecked")
-  private static final <T> scala.collection.immutable.List<T> nil() {
+  private static <T> scala.collection.immutable.List<T> nil() {
     return (scala.collection.immutable.List<T>) scala.collection.immutable.Nil$.MODULE$;
   }
 
@@ -1425,7 +1344,7 @@ public final class Main {
 
   private void launch(Options options, CommandLine line, Level logLevel, boolean logMasking) {
     if (line.hasOption(LANGUAGE_SERVER_OPTION)) {
-      runLanguageServer(line, logLevel);
+      LanguageServerApi.launchLanguageServer(line, logLevel);
     } else {
       try {
         var conf = parseProfilingConfig(line);
